[メソッドの作成]フィールドクラスに、キャラクター移動に伴うフィールドを書き換えるメソッドを追加
