+++ Main/src/character/move/Action.java
@@ -1,0 +1,8 @@
+package character.move;
+
+import point.Point;
+
+public interface Action {
+    Point go(Point point);
+    Point former(Point point);
+}
