+++ Main/src/character/move/Left.java
@@ -1,12 +1,11 @@
-package field.point.move;
+package character.move;
 
-import field.point.Point;
+import point.Point;
 
-class Up implements Action{
-
+class Left implements Action{
     @Override
     public Point go(Point point) {
-        return Point.init(point.getRow() + MoveLength.normal, point.getColumn());
+        return Point.init(point.getRow(), point.getColumn() - MoveLength.normal);
     }
 
     @Override
