+++ Main/src/character/move/MoveLength.java
@@ -1,4 +1,4 @@
-package field.point.move;
+package character.move;
 
 class MoveLength {
     static final int normal = 1;
