+++ Main/src/character/move/Up.java
@@ -1,11 +1,12 @@
-package field.point.move;
+package character.move;
 
-import field.point.Point;
+import point.Point;
 
-class Right implements Action{
+class Up implements Action{
+
     @Override
     public Point go(Point point) {
-        return Point.init(point.getRow(), point.getColumn() + MoveLength.normal);
+        return Point.init(point.getRow() + MoveLength.normal, point.getColumn());
     }
 
     @Override
