+++ Main/src/field/Field.java
@@ -1,19 +1,26 @@
 package field;
 
 import character.Character;
+import character.move.Direction;
 import field.obstacle.Obstacle;
-import field.point.move.Move;
+import point.Point;
 
 public class Field {
     private String[][] map = new String[Range.MAX_ROW][Range.MAX_COLUMN];
+    private static final String DOT = ".";
 
     private Field(){
         for(int i =0; Range.rowIsValid(i); i++){
             for(int j = 0; Range.columnIsValid(j); j++){
-                map[i][j] = ".";
+                map[i][j] = DOT;
             }
         }
     }
+
+    public static Field create(){
+        return new Field();
+    }
+
 
 
     public void obstacleInit(Obstacle obstacle){
@@ -25,7 +32,17 @@ public class Field {
         this.map[character.pointIs().getRow()][character.pointIs().getColumn()] = character.toString();
     }
 
-    public void move(Move move){
-
+    public void normalPoint(Point point){
+        this.map[point.getRow()][point.getColumn()] = DOT;
     }
+
+    public void charPoint(Character character){
+        this.map[character.pointIs().getRow()][character.pointIs().getColumn()] = character.toString();
+    }
+
+    public void charMovement(Character character, Direction direction){
+        this.map[character.pointIs().getRow()][character.pointIs().getColumn()] = DOT;
+        this.map[character.move(direction).getRow()][character.move(direction).getColumn()] = character.toString();
+    }
+
 }
