--- Main/src/field/point/move/Action.java
@@ -1,8 +1,0 @@
-package field.point.move;
-
-import field.point.Point;
-
-interface Action {
-    Point go(Point point);
-    Point former(Point point);
-}
